Move CubismDataSourceStatic into main dist
